refactor(broker): step 2/6 — extract ReplayDiscrepancy types into discrepancy.ts
Step 2 of the directory split. Extracts the discriminated union and
the two in-memory state shapes into a dedicated types module:

  cost-ledger/replay-check/
  └── discrepancy.ts  (~374 LOC: ReplayDiscrepancy + ReplayCheckReport
                       + ReplayedBudget + ReplayedCrossing)

This is the largest single block in replay-check.ts (~320 LOC of pure
discriminated-union variants). Extracting it:

- Lets future consumers (Go/Rust SDK doc generators, dashboards)
  import the wire-shape catalogue from one well-known module.
- Decouples the orchestrator's runtime logic from the discrepancy
  taxonomy.
- Drops replay-check.ts from 1485 LOC to 1150 LOC.

replay-check.ts re-exports `ReplayCheckReport`, `ReplayedBudget`, and
`ReplayDiscrepancy` so external consumers and the package's index
file see no change. `ReplayedCrossing` stays internal (only the
orchestrator's scan loop builds it).

All 249 broker tests pass; typecheck clean.
